agent-worker/requirements.txt
# Lead Friendly AI Voice Agent Worker
# LiveKit Agents framework + plugins

livekit>=0.13.0
livekit-agents>=0.10.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-elevenlabs>=0.8.0
livekit-plugins-anthropic>=0.4.0
livekit-plugins-silero>=0.7.0

# HTTP client for API callbacks
httpx>=0.27.0

# Environment variable loading
python-dotenv>=1.0.0
